Move upload serializers to own file
