+++ app/stac_api/management/commands/list_asset_uploads.py
@@ -6,7 +6,7 @@ from django.core.serializers.json import DjangoJSONEncoder
 
 from stac_api.models import AssetUpload
 from stac_api.s3_multipart_upload import MultipartUpload
-from stac_api.serializers.serializers import AssetUploadSerializer
+from stac_api.serializers.upload import AssetUploadSerializer
 from stac_api.utils import CommandHandler
 from stac_api.utils import get_asset_path
